Update index.js file
